Separated the Individual class into its own file to support testing and modularity
